Storage of a list of seqnum->ledger mappings in zookeeper.
The path in zookeeper is based on the URI.
A bookkeeper URI is defined as bookkeeper:/base/path/t:tableName/r:region/seqNum
This maps to a znode /base/path/t:tableName/r:region, seqNum defines an entry in the data of the znode. The data is serialized in protobuf form.

I added a method to add to the list and to read the list. Adding to the list ensures that every entry in the list has been returned by #listLogs before updating. This avoids splitbrain (see LatestRegionZNodeVersion).